Split anteHandleMessage into multiple messages
