+++ utility/unit_of_work/tx_message_handler.go
@@ -219,68 +219,6 @@ func (u *baseUtilityUnitOfWork) handleMessageChangeParameter(message *typesUtil.
 	return u.updateParam(message.ParameterKey, v)
 }
 
-// REFACTOR: This can be moved over into utility/types/message.go
-func (u *baseUtilityUnitOfWork) getSignerCandidates(msg typesUtil.Message) ([][]byte, coreTypes.Error) {
-	switch x := msg.(type) {
-	case *typesUtil.MessageSend:
-		return u.getMessageSendSignerCandidates(x)
-	case *typesUtil.MessageStake:
-		return u.getMessageStakeSignerCandidates(x)
-	case *typesUtil.MessageUnstake:
-		return u.getMessageUnstakeSignerCandidates(x)
-	case *typesUtil.MessageUnpause:
-		return u.getMessageUnpauseSignerCandidates(x)
-	case *typesUtil.MessageChangeParameter:
-		return u.getMessageChangeParameterSignerCandidates(x)
-	default:
-		return nil, coreTypes.ErrUnknownMessage(x)
-	}
-}
-
-func (u *baseUtilityUnitOfWork) getMessageStakeSignerCandidates(msg *typesUtil.MessageStake) ([][]byte, coreTypes.Error) {
-	pk, er := crypto.NewPublicKeyFromBytes(msg.PublicKey)
-	if er != nil {
-		return nil, coreTypes.ErrNewPublicKeyFromBytes(er)
-	}
-	candidates := make([][]byte, 0)
-	candidates = append(candidates, msg.OutputAddress, pk.Address())
-	return candidates, nil
-}
-
-func (u *baseUtilityUnitOfWork) getMessageEditStakeSignerCandidates(msg *typesUtil.MessageEditStake) ([][]byte, coreTypes.Error) {
-	output, err := u.getActorOutputAddress(msg.ActorType, msg.Address)
-	if err != nil {
-		return nil, err
-	}
-	candidates := make([][]byte, 0)
-	candidates = append(candidates, output, msg.Address)
-	return candidates, nil
-}
-
-func (u *baseUtilityUnitOfWork) getMessageUnstakeSignerCandidates(msg *typesUtil.MessageUnstake) ([][]byte, coreTypes.Error) {
-	output, err := u.getActorOutputAddress(msg.ActorType, msg.Address)
-	if err != nil {
-		return nil, err
-	}
-	candidates := make([][]byte, 0)
-	candidates = append(candidates, output, msg.Address)
-	return candidates, nil
-}
-
-func (u *baseUtilityUnitOfWork) getMessageUnpauseSignerCandidates(msg *typesUtil.MessageUnpause) ([][]byte, coreTypes.Error) {
-	output, err := u.getActorOutputAddress(msg.ActorType, msg.Address)
-	if err != nil {
-		return nil, err
-	}
-	candidates := make([][]byte, 0)
-	candidates = append(candidates, output, msg.Address)
-	return candidates, nil
-}
-
-func (u *baseUtilityUnitOfWork) getMessageSendSignerCandidates(msg *typesUtil.MessageSend) ([][]byte, coreTypes.Error) {
-	return [][]byte{msg.FromAddress}, nil
-}
-
 func (u *baseUtilityUnitOfWork) checkBelowMaxChains(actorType coreTypes.ActorType, chains []string) coreTypes.Error {
 	// validators don't have chains field
 	if actorType == coreTypes.ActorType_ACTOR_TYPE_VAL {
